requirements.txt
langchain
pinecone_client==2.2.2
streamlit==1.26.0
unstructured
unstructured[pdf]
openai
chromadb
tiktoken